refactor(runner): split dns module by responsibility
